zloop timer creation fix.
When adding a zloop timer AFTER the loop had been started - and especially when doing it from a timeout handler - the timer did not have a valid 'when' value and would therefore make the loop never sleep/poll but spin at max speed instead.

This fix simply sets a default value of -1 to the 'when' attribute of a timer and then sets it to a correct value just before using it.