Deal with zmalloc failing to allocate memory in the zframe module.
Check the return of zmalloc in zframe_new().  Deal with the cases
where this function is called elsewhere in the code (and corresponding
changes related to those changes).

Refactor zmsg_addstr()/zmsg_pushstr() to use a common base function.

Currently, zframe_send() does *not* destroy the frame on error, even
if ZFRAME_REUSE is not set.

There is a FIXME comment in the zmsg module for zmsg_load().  It
currently just breaks out of the loop like all other error conditions
if zmsg_add() fails, but this seems like it could cause silent data
corruption (in all cases), so it should be looked at when going
through the zmsg module later.

Signed-off-by: AJ Lewis <aj.lewis@quantum.com>