Refactor zmsg_load() to enable error handling
zmsg_load() used to just break out of the fread loop and return a
zmsg_t structure with the frames it was able to read, but not
returning any error status back to the calling function.  This changes
the API to return an error code from the function, and makes the
return message pointer an argument to zmsg_load().

On error, zmsg_load() now rewinds the state of the zmsg_t passed in.
Any frames added to the message are destroyed, and if the message was
created by zmsg_load() it is destroyed as well.

Signed-off-by: AJ Lewis <aj.lewis@quantum.com>