Deal with zmalloc failing to allocate memory in the zhash module.
Check the return of zmalloc in zsocket_new().  Deal with the cases
where this function is called elsewhere in the code.

Refactor s_item_insert() to allow errors to be propagated back.  Check
the return value of zmalloc and return an error if it fails.  Adjust
callers accordingly.

Adjust zhash_insert to detect zmalloc error and s_item_insert errors.
Now instead of returning either -1 if the key is present and 0 on
success, it can also return a positive error code on failure.

Signed-off-by: AJ Lewis <aj.lewis@quantum.com>