Make zctx_interrupted volatile
To be absolutely safe, in the presence of signals a flag variable
should use the volatile keyword. If the compiler was capable of
inter-library optimization, then it would be possible for it to
assume that since the interrupted flag is never changed by code
in the normal path that it was always true.