Zero-copy methods slated for removal
There are multiple problems with these methods but mainly the accumulation
of technical debt for uncertain benefit. It makes the CZMQ API complex and
is arguably a case of premature optimization. Real zero-copy use cases are
very rare and they can work directly with the zmq_msg API. The zero-copy
semantics have also leaked in confusing ways (e.g. zframe_freefn). Some of
the methods copied data in any case.

If we really want a zero copy API in CZMQ we should put this into a
single znocopy module with clear semantics, and we should design a nicer
zero-copy API.