Added zconfig module
This saves and loads ZPL format text files. We use this already in FileMQ
for configuration, and I want to use it for CurveZMQ certificates as well.
The alternative is to use a flat file format like RFC 4716, but that does
not allow us to store multiple keys in one file, which would be useful
for authentication (e.g. server holding client long-term public keys in a
single file). ZPL is useful for any kind of configuration, so adding this
to CZMQ seems useful.

Note that this code originated in libzfl, and has lived for a while in
FileMQ (which will now use the zconfig class instead).