Add an "add" function for zpoller.
Added a function to add a reader after the initial poller has been
created. I'm not sure how correct this patch is but it passes the self
test. I figured that because both the use case and the patch still
conforms to the zpoller description that it'll be okay.

Also can I get discussion on the documentation and function name for
this?

Use Case:
Creating sockets in a loop and adding them to a poller after it's been
created.

Limitations:
  * Appends a reader so it'll have the lowest priority.
  * There's no corresponding remove function.

Signed-off-by: Shannen Saez <shannenlaptop@gmail.com>