Use system uuid functions in zuuid
On OS X, libuuid is part of the base system, however isn't found by
autoconf as might be expected. Consequently this commit adds an explicit
check for OS X and defines the HAS_LIBUUID symbol if that is what we are
running. This makes zuuid_new() significantly faster on OS X.