Problem: need gossip discovery service
This is to provide an alternative or complement to zbeacon for networks
with one or more static services. The gossip service will run on every
node, so that nodes maintain and synchronize their maps of the network.
This will allow any mix of static and dynamic nodes on a network, where
static nodes bind, and dynamic nodes connect.

The main use case for zgossip is Zyre, which builds a fully-connected
cluster across all known nodes in a network.

This commit is for the basic zgossip API, which is built on the zactor
class. The code doesn't do anything real yet.