Problem: zmq_socket returns void *, not zsock_t
zsys_socket was inconsistently using socket and handle, the code
would not build on ZMQ v2 for example. Since 'socket' is dangerous
to use (it silently resolves to the library function if there's no
variable in scope), I've switched to using 'handle' in this code.

Also, the socket configuration API is zsocket_set_xxx, which may be
replaced by direct calls to libzmq when we deprecate zsockopt.c.