Problem: zhash_foreach() is clumsy and horrible
Callbacks are nasty, and I much prefer using the zlist first/next
style iterators, so I've added this to zhash and it works nicely.

Solution: use simple iterator API as learned from zlist, so caller
can get items from hash table in some arbitrary order.

Difficulty: we need to provide both key and value, so the methods
return the value, and a helper method (zhash_cursor) to return
the key.