Problem: zmsg_print and zframe_print go to stdout
All logging has to go to a single controllable destination, which
may be stdout or stderr, or may be the system logging facility.

Solution: fix zmsg_print and zframe_print to use zsys_debug for
their output. Deprecate zmsg_fprint and zframe_fprint, as it is
cleaner to call zsys_set_logstream (somefile) to direct all log
traffic to the desired file.