Problem: was passing wrong size to zmq_getsockopt on keys
libzmq requires receiving buffer to be exactly 32, or 41 bytes
for a binary or Z85 key respectively. CZMQ was passing a 255
byte buffer, and then ignoring the EINVAL return code. This was
caught by the new --with-militant assert checking (yay!)

Solution: pass a 41 byte buffer as required by the zmq_getsockopt
man page.