Problem: unreliable dependency on libsodium
CZMQ doesn't need libsodium, as all relevant calls are wrapped in the
zmq_util layer. Furthermore, detecting and using libsodium in CZMQ is
fragile as it is too easy to build libzmq and CZMQ inconsistently.

Additionally, the previous approach assumed CZMQ would be rebuilt if
there was a change to libzmq.

Solution: remove all dependencies on libsodium. Use zsys_has_curve()
to detect and report whether libzmq was built with libsodium or not.
This lets us switch CURVE on/off dynamically by installing different
builds of libzmq.