Problem: zgossip fails with large swarms / set sizes
I've extended the test_zgossip tool to allow tests with larger swarms
and set sizes, and have tested swarms of up to 1,000 nodes and set sizes
of 8K or higher.

Two issues hit. One is system limits on file handles. On Linux this is
1024 by default, and the hard upper limit seems to be 4096. Each node
in a swarm needs 4 file handles, and we use 8 for infrastructure, so
1024 handles lets us run 254 nodes, the default for the test case.

Second issue is pipe overflow with large set sizes. There are two places
this hits. First, the DEALER socket that each peer uses to talk to a
remote. When this is full, the node simply blocks and when enough nodes
block, the test case deadlocks. Second, the ROUTER socket that each peer
uses to talk to its client peers. When this is full, the node drops
outgoing PUBLISH messages, and the test case never completes.

The solution in both cases is to set an infinite HWM. The risk is that
when/if applications use very large set sizes, they will cause memory
overflow. The soltion, if this problem ever hits, is to use a CBFC model
on tuple publishing.