Problem: not working properly on ZeroMQ/2.x any more
The problems are mainly with inproc bind/connect. In 2.x these have
to come in the right order, and with multiple threads, we need a
short pause before connect is safe to go.

I might fix this in zsock_connect by pause/looping a few times, if
we're on 2.x and the connect fails.