move a line back to the right spot; make sure all sockets get destroyed
I moved the call to zsocket_test up ahead of the gossip call, because
I was using an old version of libzmq; I upgraded and am moving it back,
gossip's test runs OK now.

I was getting a couple of leaked socket message from the test. These
are corrected now.