Make the self-tests complete OK for libzmq v2.
I was able to reproduce the problem from the information gleaned from the
autmated tests reported from github. As it turns out, zmq before 3.2.0,
just returned -1 from the unbind command. So, I added #ifs to skip the
unbind calls if the libzmq stuff is old enough.

I also lowered the retry max from 30 to 10. 30 just seemed, after some
thinking, a bit overkill, and inefficient.

The zinterface.txt thing is a mystery to me, mayhaps a leftover from
my updating the source to the current version of czmq. Someday I'll
understand.