Problem: trex code is flaky
In real use this code breaks; it does not, for instance, return
capture groups correctly. I've switched to another similar piece
of code, SLRE, which was MIT-licensed for v1.0 (has since gone
to GPLv2 which isn't useful for us).

Solution: drop trex code and replace with slre v1.0.

FWIW I also extended the regexp syntax to support \w, \W, \D, \a,
and \A.