Problem: bind code was inelegant
Solution: use zrex for parsing. Combine random and sequential port
scanning into single loop, as they're basically the same work. I've
not tested this to exhaustion.

Note: we can now trivially add randomizaton for other endpoints, and
it might be smarter to return the last bound endpoint as a property
rather than via the return code.