Problem: sending complex commands to actors is too clumsy
C isn't a neat language for sending mixed types as messages, e.g.
a string, an int, and a binary chunk, as one message. In zbeacon
this makes for a cumbersome API.

Solution: use a 'picture' and va_args to allow arbitrary types in
the send command. Implement this in zsock_send and zsock_recv,
which were waiting for something fun.