Problem: zlist got out of control
This was always meant to be a *simple* list container, otherwise
why would we go to all the pain of using a singly-linked list
rather than the obviously easier double-linked list... yet after
two days of refining the API while keeping the old API intact,
this thing isn't finished.

To make this worse, zlist is copying a lot from zring, except it
can't do it properly, lacking the backwards links.

Solution: simple is less, not more.

1. revert all changes to zlist and keep it as it was a week ago.

2. zring is a better place to (carefully) add new semantics.