Problem: zhash public API was broken
Changing keys from 'const char *' to 'void *' broke existing applications,
and isn't necessary. The majority use case is for char keys, so the API
can show this. The marginal case is for void * keys, which zhash does
internally without modifying the API. Apps that use non-string keys will
have to cast their keys to 'const char *', which is fine (the alternative
is that all existing apps have to add a cast just to allow a few unusual
apps to not add one.)