Disable distcleancheck due to the dist generated files
The automated generation of the ChangeLog file introduced an issue in
the distribution chain, leading to ChangeLog being left at distclean
time, while it should have been removed. The distcleancheck_listfiles
workaround (commented) works fine on modern systems, but may cause
issues on older unix systems. This commit disables distcleancheck, and
leaves to CZMQ maintainers the choice to enable distcleancheck_listfiles
or not