Problem: node insert/delete code is clumsy
In fact a doubly-linked list is easy to work with if you always initialize
a new node to point to itself. Then, inserting and removing a node is the
same operation (swap pointers with prev/next nodes). As we're doing this in
a few places, and it's the most tricky area of such code, it's worth doing
right.

Solution: do it right using a s_node_relink method.

Problem #2: inner loop check on self->comparator is clumsy and unnecessary.

Solution: define a default comparator so this test can disappear.