Problem: confusing delayed compute of timer 'when'
There is no functional reason to delay the calculation of the when
property for new timers. However this makes the code more complex
than it has to be.

Solution: calculate the 'when' on new timers, and simplify the rest
of the code a little.

Note: this also lets us insert timers correctly into a sorted list.