Problem: zhash got messy
The Features added over the last months made the API complex
and nasty (void * for keys, really?) and broke compatibility.

Solution: revert zhash to its v2 simpler state, which can be
deprecated or perhaps kept as a "simple hash" container. All
the nasty stuff should be kept for zhashx.

Also switched to using zhashx internally in CZMQ (not in places
where it affects the API).