PROBLEM: zring was removed, but can't be implemented using zlistx and zhashx
zring has been removed recently because it's a weird data structure.
Yet, I have found it really useful, so tried to implement it on top
of zlistx and and zhashx, by storing list handles in the hash.
However, when deleting an item from the hash, one needs to convert a
list handle to the item stored in it, which is currently not possible,
because the pointer is opaque.

SOLUTION: add a function to convert a handle into the associated item.