Dir patch does not duplicate delete patches
Problem: The zdir_patch object does not duplicate a patch that is
a delete operation. The duplication method requires that the file's
digest is not NULL, but a delete operation will not have a file to
read to calculate the digest.

Solution: Added a test for the operation type before the zdir_patch
object is destroyed for having an empty digest.