Problem: zbeacon does not set default interval
The pre-actor version used a default of 1000 msec if the caller did
not specify an interval. The actor version uses no default and if
the caller sends an interval of zero (as Zyre did), it sends UDP
beacons at full speed, killing the network.

Solution: zbeacon should treat interval=0 as meaning "default".