PROBLEM: can't shut down an actor system in a controlled way after interrupt
After receiving an interrupt, all zloops and zpollers are put into a
state where they refuse to start/continue running. Unfortunately, this
makes it impossible to terminate a system of zactors in a crontrolled
order.

As an example, assume there is one actor used to linearize and format
output of all other actors. We definitely want to shut down this actor
after all others, so that we are able to log all events which occur
during systen shutdown. This doesn't work at the moment when the using
the default signal handling approach.

The patch adds two methods which can be used to make zloops and
zpollers ignore the zsys_interrupted global.

   void zloop_ignore_interrupts(zloop_t *self);
   void zpoller_ignore_interrupts(zpoller_t *self);

Note that there were two additional places in the code base which take the
global zsys_interrupted into account and had to be changed to always
ignore zsys_interrupted. This code was vulnerable to race conditions
when setting zsys_interrupted and could leak to programs which would
not be interruptible.