Problem: No way to run verbose tests via make
Solution: Make check-verbose.

Rationalle: Newer versions of automake >= 1.13 force `make check` to be
run through a parallel test runner which turns off all output. So we
need a way to at least see the non-verbose output in that sort of
environment, hence check-local.

check-verbose comes about because when CI failures or hangs occur, there
is very little information to go on, so making it always dump as much
information as possible should hopefully help diagnose these issues in
future.