Problem: zarmour API changes depending on libzmq version
The internal and external APIs shift around depending on Z85
support. This seems fragile. We could for instance add Z85
encoding/decoding into zarmour if libzmq doesn't support it.

Solution: always allow Z85 to be chosen by the API and return
null on encode/decode if it isn't supported.

This also makes the code simpler.