Problem: inconsistent use of fresh/constant in APIs
E.g. sometimes fresh = "1" is used instead of "constant = "0" in
arguments. This is not clean; fresh is meant for return values.

Also, some arguments have 'constant = "1"' which is already the
default for strings and buffers. Specifying this explicitly is
misleading.

Solution: fix misuse of 'fresh' and remove default constant = "1"
attributes on strings, buffers, and formats.