Problem: JNI binding was not using proper types for classes
Solution: a lot of hacking on the JNI generator...

Fixes #1212

This code has to be refactored at some stage; it is nasty. Every
data type has different handling in different places.

We also have hit some semantic walls in the API grammar that we
need to fix, e.g. we don't know in some cases where the size of
a buffer is, to return from C to Java.