Problem: does not build on MSVC
Solution: ... a lot of fixes to the zproject code generators. The main
ones are adding libsodium as dependency; copying the import files from
sibling directories (see #395), generating all the right files; fixing
a whole lot of property references; fixing a lot of backslash escaping
that had gotten lost.

I've tested this on CZMQ so far, for VS2013. I'm unable to test VS2008
and I assume it's still broken. VS2010, VS2012 and VS2015 use the same
backend as VS2013 and so presumably work.