Problem: codebase is filled with useless OOM checks
The code now asserts when it cannot allocate an object.

Note that OOM checking made a lot of code more complex and was not
working, as ZeroMQ itself stops working when the heap is exhausted.

Also, if a programmer is creating unbounded architectures, one
cannot really expect them to deal with subtle error conditions
properly either.

Solution: remove the OOM checks.

Fixes #1193