Problem: zauth sometimes hangs in self test
Analysis: test code was wrongly assuming it was always safe to send
on a socket after 'connect'. In fact if the connect fails due to an
authentication failure, there is a small window between calling the
'connect' and the peer rejecting the connection. In this window, a
send will work. After the window, it will hang forever.

Solution: set send timeout on test socket.