Problem: status attribute defined in project.xml
The properties of a thing should be defined in that thing. E.g.
where would be specify the version of a class?

This means the state information should be inside the XML file,
not in project.xml.

Also, I'd recommend "state" and not "status".

Solution: remove these definitions from project.xml.