Problem: use_fd needs to be used manually by apps
Solution: add new zsys_auto_use_fd option that enables a smarter
behaviour. If czmq is built with --with-libsystemd and
zsys_auto_use_fd is enabled, before binding a new socket the system
API will be called to check if a socket with a matching path (if IPC)
or port (if TCP) already exists and is managed by the system. If so,
zmq_use_fd will be automatically set on the socket.
The first implementation supports systemd's API, but other systems
like upstart or inetd could be added as well.