Problem: zlist_sort does not work according to docs
The docs say this does an ASCII sort but in fact it does not,
and it won't work unless you pass a comparison function. This
is over-design for 90% of cases.

Solution: extend API slightly so that a NULL comparison function
gets treated as "strcmp", giving the desired effect for most
simple cases.