Problem: zsock options could be available at build time but not at runtime
Solution: rework the sockopts compatibility layer to generate a
runtime check together with the build time check. In case the runtime
check fails, print an error stating what the minimum required is vs
the runtime version. It could be argued that an assert should be used
instead, but this follows the same principle as the existing
implementation where options not available at buildtime are compiled
to simply return 0 and be no-ops, rather than asserting.
Rework the sockopts.xml layout to let the options of each zmq version
be grouped together in different and independent blocks to allow mix
and match. Introduce a major_removed element to indicate in which
version an option has been removed, to allow a finer granularity in
the build time and run time checks.