Problem: ziflist does not support IPv6
Solution: use getnameinfo in s_interface_new to get the string format
of the interface address and parameters.
Add new (draft) methods ziflist_new_ipv6 and ziflist_reload_ipv6 that
also look for IPv6 addresses in order to maintain backward
compatibility.
Only implemented for operating systems that support getifaddrs, so
not with ioctl calls (they support only IPv4) and not on Windows
(help needed to implement it).