Problem: Travis rebuilds libsodium every time
Solution: use packages on Ubuntu and brews on OSX. The packages and
the brews are always kept up to date, so it's no use to rebuild the
libsodium stable branch manually everytime.
Given apt-get is already ran for all builds, adding the OBS repo to
the sources and always installing libsodium is a negligible
overhead.