Problem: src addr of v6 beacons is random garbage
Solution: use two UDP sockets, one for sending and one for receiving.
For IPv6 we need two sockets. At least on Linux, IPv6 multicast packets
are NOT received despite joining the group and setting the interface
option UNLESS the socket is bound to in6_addrany, which means the kernel
will select an arbitrary IP address as the source when sending beacons
out. This breaks zbeacon as the protocol uses the source address of a
beacon to find the endpoint of a peer, which is then random and
useless (could even be associated with a different interface, eg: a
virtual bridge).
As a workaround, use a different socket to send packets. So the socket
that receives can be bound to in6_addrany, and the socket that sends
can be bound to the actual intended host address.