Problem: zsys_shutdown does not call zmq_term when it closes sockets
Solution: decrement count of open sockets if it manages to close them
from the internal list.
The total count of open sockets might go up without a socket being
appended to the list, but if a socket is appended to the list then
count is always incremented. So it's safe to decrement the count as
sockets are popped out of the list and closed.
This increases the chances that a clean exit is achieved even with
dangling sockets.