Problem: getnameinfo does not append iface name on Solaris
Solution: if the address is IPv6 linklocal and it does not have the
interface name appended with % do it manually in ziflist and
zsys_udp_recv.
On Linux getnameinfo does it, on Solaris it does not.
Link local addresses cannot be used without the interface, as they
are not guaranteed to be unique behind the link on which they
are configured.