Problem: Sending to INADDR_BROADCAST on Windows only results in a single packet being sent over the interface with the highest metric.
This results in nodes not being seen on all interfaces when using the any ("*") interface, metric can be observed using the command 'route print'

Solution: SIO_GET_INTERFACE_LIST offers a cheap way of enumerating interfaces available to a specific socket so a packet can be sent per interface. Profiling shows this method at 0.006ms per invokation against 0.5ms for GetIpAddrTable meaning that this can be used on every beacon send instead of caching the results.