Problem: parsing commandline is hard
Solution: add ligthweight and minimal API for accessing argv in the most
ergonomic, yet very flexible, way. Class zargs_t distinguish between
(positional) arguments and (named) parameters. And provide different
czmq compatible API for accessing those.

It aims to be getopt compatible, so comes with -- as a way to
distinguish between arguments and parameters. In future it will handle
the --with=value and -Wvalue variants as well.