problem: zframe_meta documentation is incorrect
zframe_meta returns the value from zmq_msg_gets directly and according
to http://api.zeromq.org/4-2:zmq-msg-gets

The zmq_msg_gets() function shall return the string value for the
property if successful. Otherwise it shall return NULL and set errno to
one of the values defined below. The caller shall not modify or free the
returned value, which shall be owned by the message. The encoding of the
property and value shall be UTF8.