problem: User-Id field in zap response is not populated.
According to ZAP spec:

user id: this MAY provide the user identity in case of a 200 status, for
use by applications. For other statuses, it SHALL be empty.

Solution:  This adds a char pointer in the zap request struct that can
point to the the username, client_key, or principal fields.

With this change zframe_meta(frame, "User-Id") returns the public key
for a connection.  This enables a use case where you have a long lived
connection authenticated by a key or password and want to verify that the
user/key still exists.

    zframe_t *frame = zframe_recv(server);
    const char *user_id = zframe_meta(frame, "User-Id");
    if(user_id!=NULL) {
        zcert_t *c = zcertstore_lookup(store, user_id);
        if(!c)
            //Certificate no longer exists in store!
    }