Problem: option state code is draft focused
It currently only honors "draft" and would not allow
`state = "deprecated"` to be propagated into api/zsock_option.api.

Solution: If the state attribute is set for an option in
src/sockopts.xml, also put it in the corresponding method in
api/zsock_option.api, whatever its value is.

Currently this has no effect as all options are stable, but it will as
soon as an option is deprecated.