Problem: tests hang due to ZMQ_POLL_MSEC used with zpoller and zmq_poller
Solution: remove it. It is a backward-compat macro to use only with
the zmq_poll API, which used to take nanoseconds until libzmq 3.x
where it was changed to milliseconds.
In the zbeacon test it was used with zpoller_wait, which itself used it,
which means that when building with libzmq 2.x the timeout was twice
multiplied by 1000 and caused the test to hang.
Also remove it from the zmq_poller usage, which is new and has always
taken milliseconds.

Fixes #1825